+++ packages/remix-fastify/src/plugin.ts
@@ -41,7 +41,7 @@ interface PluginOptions {
   getLoadContext?: GetLoadContextFunction;
   /**
    * purge the require cache in development when not using the new dev server
-   * @default process.env.NODE_ENV === "development"
+   * @default false
    */
   purgeRequireCacheInDevelopment?: boolean;
   /**
@@ -69,21 +69,17 @@ async function loadBuild(build: ServerBuild | string): Promise<ServerBuild> {
 
 let remixFastify: FastifyPluginAsync<PluginOptions> = async (
   fastify,
-  options = {}
+  options = {},
 ) => {
   let {
     build,
     mode = process.env.NODE_ENV,
     rootDir = process.cwd(),
-    purgeRequireCacheInDevelopment = process.env.NODE_ENV === "development",
-    unstable_earlyHints: earlyHints,
+    purgeRequireCacheInDevelopment = false,
+    unstable_earlyHints: earlyHints = true,
   } = options;
-  invariant(build, "You must provide a build");
+  invariant(build, "you must pass a remix build to the plugin");
   let serverBuild: ServerBuild = await loadBuild(build);
-
-  if (mode === "development" && !!serverBuild.dev) {
-    purgeRequireCacheInDevelopment = false;
-  }
 
   if (!fastify.hasContentTypeParser("*")) {
     fastify.addContentTypeParser("*", (_request, payload, done) => {
@@ -153,14 +149,13 @@ let remixFastify: FastifyPluginAsync<PluginOptions> = async (
 
   if (mode === "development" && typeof build === "string") {
     fastify.all("*", async (request, reply) => {
-      invariant(build, "we lost the build");
+      invariant(build, "you must pass a remix build to the plugin");
       invariant(
         typeof build === "string",
-        `to support "HMR" you must pass a path to the build`
+        `to support "HMR" you must pass a path to the build`,
       );
-      if (purgeRequireCacheInDevelopment) {
-        purgeRequireCache(build);
-      }
+
+      if (purgeRequireCacheInDevelopment) purgeRequireCache(build);
 
       let loaded = await loadBuild(build);
